server/action-library: Add Outreach mirror actions
Change-type: minor
