+++ .circleci/config.yml
@@ -246,11 +246,11 @@ jobs:
 
       - run:
           name: End To End Tests
-          command: make test-e2e-sdk test-e2e-server COVERAGE=1 INTEGRATION_GITHUB_TOKEN=$GITHUB_TOKEN INTEGRATION_GITHUB_SIGNATURE_KEY=$GITHUB_SIGNATURE_KEY INTEGRATION_OUTREACH_APP_ID=$OUTREACH_APP_ID INTEGRATION_OUTREACH_APP_SECRET=$OUTREACH_APP_SECRET && make test-e2e-ui COVERAGE=0
+          command: make test-e2e-sdk COVERAGE=1 && make test-e2e-server COVERAGE=1 INTEGRATION_GITHUB_TOKEN=$GITHUB_TOKEN INTEGRATION_GITHUB_SIGNATURE_KEY=$GITHUB_SIGNATURE_KEY INTEGRATION_OUTREACH_APP_ID=$OUTREACH_APP_ID INTEGRATION_OUTREACH_APP_SECRET=$OUTREACH_APP_SECRET && make test-e2e-ui COVERAGE=0
 
       - run:
           name: End To End Tests (without Redis)
-          command: make test-e2e-sdk test-e2e-server COVERAGE=1 DISABLE_REDIS=1 && make test-e2e-ui COVERAGE=0 DISABLE_REDIS=1
+          command: make test-e2e-sdk COVERAGE=1 DISABLE_REDIS=1 && make test-e2e-server COVERAGE=1 DISABLE_REDIS=1 && make test-e2e-ui COVERAGE=0 DISABLE_REDIS=1
 
       # Store artifacts for debugging purposes
       - store_test_results:
