+++ apps/server/card-loader.js
@@ -72,6 +72,7 @@ module.exports = async (context, jellyfish, worker, session) => {
 		await loadCard('contrib/triggered-action-integration-github-mirror-event.json'),
 		await loadCard('contrib/triggered-action-integration-front-mirror-event.json'),
 		await loadCard('contrib/triggered-action-integration-discourse-mirror-event.json'),
+		await loadCard('contrib/triggered-action-integration-outreach-mirror-event.json'),
 		await loadCard('contrib/triggered-action-support-summary.json'),
 		await loadCard('contrib/triggered-action-support-reopen.json'),
 		await loadCard('contrib/triggered-action-support-closed-issue-reopen.json'),
