+++ lib/action-library/index.js
@@ -42,7 +42,8 @@ const mirror = async (type, session, context, card, request) => {
 		type,
 		environment.integration[type],
 		card,
-		syncContext.fromWorkerContext(type, context, request.context, session),
+		syncContext.fromWorkerContext(
+			type, context, request.context, context.privilegedSession),
 		{
 			actor: request.actor,
 			defaultUser: environment.integration.default.user,
@@ -104,6 +105,13 @@ module.exports = {
 		pre: _.noop,
 		handler: async (session, context, card, request) => {
 			return mirror('discourse', session, context, card, request)
+		}
+	},
+	'action-integration-outreach-mirror-event': {
+		card: require('./actions/action-integration-outreach-mirror-event'),
+		pre: _.noop,
+		handler: async (session, context, card, request) => {
+			return mirror('outreach', session, context, card, request)
 		}
 	},
 	'action-oauth-associate': {
