+++ lib/sync/instance.js
@@ -97,9 +97,10 @@ const getOAuthUser = async (context, provider, actor, options) => {
 
 	assert.INTERNAL(null, defaultUserCard,
 		errors.SyncNoActor, `No such actor: ${options.defaultUser}`)
-	assert.INTERNAL(null, _.has(defaultUserCard, tokenPath),
+
+	assert.USER(null, _.has(defaultUserCard, tokenPath),
 		errors.SyncOAuthError,
-		`Actor ${options.defaultUser} does not support ${provider}`)
+		`Default actor ${options.defaultUser} does not support ${provider}`)
 
 	return defaultUserCard
 }
