+++ lib/sync/integrations/outreach.js
@@ -5,6 +5,7 @@
  */
 
 const _ = require('lodash')
+const errio = require('errio')
 const Bluebird = require('bluebird')
 
 // TODO: Create an account for each company we know about
@@ -66,13 +67,33 @@ module.exports = class OutreachIntegration {
 
 		const body = getProspect(card)
 		const method = outreachUrl ? 'PATCH' : 'POST'
+		const uri = outreachUrl || `${this.baseUrl}/api/v2/prospects`
 
 		const result = await this.context.request(options.actor, {
 			method,
 			json: true,
-			uri: outreachUrl,
+			uri,
 			body
+		}).catch((error) => {
+			if (error.expected && error.name === 'SyncOAuthError') {
+				this.context.log.warn('Could not update prospect', {
+					uri,
+					method,
+					body,
+					error: errio.toObject(error, {
+						stack: true
+					})
+				})
+
+				return null
+			}
+
+			throw error
 		})
+
+		if (!result) {
+			return []
+		}
 
 		// TODO: Validate HTTP response codes. For now we leave it
 		// as it is in order to test it production without bringing
